requirements.txt
heatmapz==0.0.4
matplotlib==3.2.1
mizani==0.7.1
numpy==1.17.2
pandas==1.0.4
plotnine==0.7.0
psycopg2-binary==2.8.3
scipy==1.4.1
seaborn==0.10.1
SQLAlchemy==1.3.8
